requirements.txt
tensorflow~=2.17.0
tqdm~=4.66.4
numpy~=1.26.4
keras~=3.5.0
matplotlib~=3.9.0
pillow~=10.4.0
opencv-python~=4.10.0.84
pandas~=2.2.2
pyttsx3~=2.91
mltu~=1.2.5
scikit-learn~=1.5.1